requirements.txt
pillow
keyboard
openai
PyInstaller